# Conflicts:
#	.gitignore
#	bsp/boards/telosb/board.c
#	bsp/boards/telosb/debugpins.c
#	drivers/common/openserial.c
#	drivers/common/opentimers.c
#	openapps/SConscript
#	openapps/openapps.c
#	openstack/02a-MAClow/IEEE802154E.c
#	openstack/02a-MAClow/IEEE802154E.h
#	openstack/02a-MAClow/IEEE802154_security.c
#	openstack/02b-MAChigh/processIE.c
#	openstack/02b-MAChigh/schedule.c
#	openstack/02b-MAChigh/sf0.c
#	openstack/03b-IPv6/forwarding.c
#	openstack/SConscript
#	openstack/cross-layers/idmanager.c
#	openstack/cross-layers/openqueue.c
#	openstack/openstack.c
#	projects/common/03oos_sniffer/03oos_sniffer.c
#	projects/telosb/03oos_openwsn/03oos_openwsn.ewp

+++ bsp/boards/eldorado/radio.c
@@ -156,7 +156,7 @@ void radio_rfOn() {
    MC13192_ATTN = 1;
 }
 
-void radio_loadPacket(uint8_t* packet, uint8_t len) {
+void radio_loadPacket(uint8_t* packet, uint16_t len) {
    // load packet in TXFIFO
    radio_spiWriteTxFifo(packet,len);
 }
